destineer/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=23.5
Pillow>=10.0
psycopg2-binary>=2.9
cloudinary>=1.36
django-cloudinary-storage>=0.3.0
celery>=5.3
redis>=5.0
scikit-learn>=1.3
numpy>=1.25
pandas>=2.1
python-decouple>=3.8
drf-spectacular>=0.27
django-extensions>=3.2
gunicorn>=21.2
